finished

+++ src/components/projects/ProjectDetails.js
@@ -15,7 +15,7 @@ class ProjectDetails extends Component {
 
   getSingleProject = () => {
     const { params } = this.props.match;
-    axios.get(`http://localhost:5000/api/projects/${params.id}`)
+    axios.get(`http://localhost:5000/api/projects/${params.id}`, {withCredentials:true})
     .then( responseFromApi =>{
       const theProject = responseFromApi.data;
       this.setState(theProject);
@@ -36,7 +36,7 @@ class ProjectDetails extends Component {
 // DELETE PROJECT:
   deleteProject = (id) => {
     const { params } = this.props.match;
-    axios.delete(`http://localhost:5000/api/projects/${params.id}`)
+    axios.delete(`http://localhost:5000/api/projects/${params.id}`, {withCredentials:true})
     .then( responseFromApi =>{
         console.log(responseFromApi);
         this.props.history.push('/projects'); // !!!         
@@ -46,17 +46,29 @@ class ProjectDetails extends Component {
     })
   }
 
+  ownershipCheck = (project) => {
+    if(this.props.loggedInUser && project.owner === this.props.loggedInUser._id){
+      return (
+        <div>
+          <div>{this.renderEditForm()} </div>
+          <button onClick={() => this.deleteProject(this.state._id)}>Delete project</button>
+        </div>
+      )
+    } 
+  }
+  
   render(){
     return(
       <div>
         <h1>{this.state.title}</h1>
         <p>{this.state.description}</p>
-        <div>{this.renderEditForm()} </div>
-        <button onClick={() => this.deleteProject(this.state._id)}>Delete project</button>
+        <div >
+          {this.ownershipCheck(this.state)}
+        </div>
         <Link to={'/projects'}>Back to projects</Link>
       </div>
     )
   }
-}
+}  
 
 export default ProjectDetails;
